store codes separately
